requirements.txt
streamlit
torch
torchvision
Pillow
requests
